+++ Sources/Montage/3 Component/6 Navigations/Tab/Tab.swift
@@ -117,7 +117,7 @@ public struct Tab: View {
                         }
                     
                         if resize == .hug, let icon, let iconButtonAction {
-                            Button.IconButton(variant: .normal(size: iconSize), icon: icon) {
+                            IconButton(variant: .normal(size: iconSize), icon: icon) {
                                 iconButtonAction()
                             }
                             .padding(.trailing, horizontalPadding ? 16 : 0)
